refactor(pages): move page components to dedicated files and update imports
Restructure the application by moving page components (Landing, Forbidden, Unauthorized, NotFound) to dedicated files in the components directory. This improves code organization and maintainability by:
- Separating page logic from route files
- Centralizing shared components
- Making imports more consistent using @/ alias
- Reducing duplication of complex UI code
