+++ app/page.tsx
@@ -1,32 +1,9 @@
-"use client"
+import LandingPage from '@/components/landing/landingPage';
 
-import Link from "next/link"
-import { Button } from "@/components/ui/button"
-import { FileText } from "lucide-react"
-import Header from "@/components/header/header"
-import Hero from "@/components/landing/hero/hero"
-import FeaturesSection from "@/components/landing/features"
-import CTASection from "@/components/landing/CTA"
-import AboutSection from "@/components/landing/about-us"
-import Footer from "@/components/landing/footer"
-
-
-export default function LandingPage() {
-
+const page = () => {
   return (
-    <div className="flex min-h-screen flex-col mt-[-85px]">
-      <Header />
-
-      <main className="flex-1">
-        <Hero/>
-
-        <FeaturesSection />
-
-        <CTASection />
-
-        <AboutSection/>
-      </main>
-      <Footer/>
-    </div>
+    <LandingPage />
   )
 }
+
+export default page;
